fix format

+++ ci/do_ci.sh
@@ -162,12 +162,12 @@ function run_ci_verify () {
   mkdir -p "${OCI_TEMP_DIR}"
 
   tar xvf docker/envoy.tar -C "${OCI_TEMP_DIR}"
-  skopeo copy "oci:$OCI_TEMP_DIR" "docker-daemon:envoyproxy/envoy-dev:latest"
-  rm -rf "${OCI_TEMP_DIR}/*"
+  skopeo copy "oci:${OCI_TEMP_DIR}" "docker-daemon:envoyproxy/envoy-dev:latest"
+  rm -rf "${OCI_TEMP_DIR:?}/*"
 
   tar xvf docker/envoy-alpine.tar -C "${OCI_TEMP_DIR}"
-  skopeo copy "oci:$OCI_TEMP_DIR" "docker-daemon:envoyproxy/envoy-alpine-dev:latest"
-  rm -rf "${OCI_TEMP_DIR}"
+  skopeo copy "oci:${OCI_TEMP_DIR}" "docker-daemon:envoyproxy/envoy-alpine-dev:latest"
+  rm -rf "${OCI_TEMP_DIR:?}"
 
   docker images
   sudo apt-get update -y
